try to fix refactor
